don't do int of a float (and afterwards to float) : the digits will be truncated!
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@1650 5890f383-b09f-485f-8f59-22bc967d67fd

+++ Cachebox Core/src/CB_Core/Math/CB_RectF.java
@@ -387,12 +387,11 @@ public class CB_RectF
 
 	public static CB_RectF ScaleCenter(CB_RectF rectangle, float ScaleFactor)
 	{
-		float newWidth = (int) (rectangle.getWidth() * ScaleFactor);
-		float newHeight = (int) (rectangle.getHeight() * ScaleFactor);
+		float newWidth = rectangle.getWidth() * ScaleFactor;
+		float newHeight = rectangle.getHeight() * ScaleFactor;
 		float newX = rectangle.Pos.x + ((rectangle.getWidth() - newWidth) / 2);
 		float newY = rectangle.Pos.y + ((rectangle.getHeight() - newHeight) / 2);
 		return new CB_RectF(newX, newY, newWidth, newHeight);
-
 	}
 
 	/**
